Changed so that the peak is now measured after a guassian fit
We noticed that there was a significant variation in the peaks, because the signal plateued at the top, and minor changes would shift the focusing position of the particle 1-2 pixels. We fixed this by fitting the peak to a gaussian, andusing findpeaks on that. The new code finds peaks using the Gaussian. Can give sub-pixel peak resolution.